Revert "ver-2.0"
This reverts commit d4bbb1666cbffc977e929e6fc034bf4e90d0867f.

+++ CodevsForStudent/AI.hpp
@@ -203,7 +203,7 @@ private:
 
 		int average = 0;
 		for (const auto& v : blockTop) average += v;
-		average /= (int)blockTop.size();
+		average /= blockTop.size();
 
 		//R’J‚ÌŒŸõEŒ`‚Ì•]‰¿
 		int flatScore = 0;
